// frontend/src/api/model/feedbackItemDto.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * YaoYao Dinner API
 * Restaurant management API
 * OpenAPI spec version: 1.0
 */
import type { FeedbackReactionDto } from './feedbackReactionDto';

export interface FeedbackItemDto {
  id: string;
  /** @nullable */
  by: string | null;
  /** @nullable */
  content: string | null;
  /** @nullable */
  eventId: string | null;
  reactions: FeedbackReactionDto[];
  reactionTotal: number;
  createdAt: string;
  updatedAt: string;
}