// frontend/src/api/model/getFeedbackSort.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * YaoYao Dinner API
 * Restaurant management API
 * OpenAPI spec version: 1.0
 */

export type GetFeedbackSort = typeof GetFeedbackSort[keyof typeof GetFeedbackSort];


export const GetFeedbackSort = {
  recent: 'recent',
  top: 'top',
} as const;